+++ packages/common/src/Components/Tabs.jsx
@@ -9,12 +9,13 @@ const tabPropTypes = {
   value: PropTypes.string.isRequired,
   className: PropTypes.string,
   children: PropTypes.node.isRequired,
+  onTabSelected: PropTypes.func,
 };
 
 export class Tab extends Component {
   static propTypes = {
     ...tabPropTypes,
-    label: PropTypes.string.isRequired,
+    label: PropTypes.oneOfType([PropTypes.string, PropTypes.object]).isRequired,
     selected: PropTypes.bool,
   };
 
@@ -46,6 +47,12 @@ export class Tabs extends Component {
     this.state = { activeTab: props.value };
   }
 
+  componentWillReceiveProps = nextProps => {
+    if (nextProps.value !== this.state.activeTab) {
+      this.setState({ activeTab: nextProps.value });
+    }
+  };
+
   getTabHeaders = tabs =>
     React.Children.map(tabs, tab => ({ label: tab.props.label, value: tab.props.value }));
 
@@ -59,7 +66,6 @@ export class Tabs extends Component {
   render() {
     const { styles, props } = this;
     const headers = this.getTabHeaders(props.children);
-
     return (
       <div role="tablist" className={styles.tabs} aria-orientation="horizontal">
         <div className={styles.tabs_headers}>
@@ -81,6 +87,9 @@ export class Tabs extends Component {
                 aria-selected={isSelected}
                 onClick={() => {
                   this.setState({ activeTab: value });
+                  if (this.props.onTabSelected) {
+                    this.props.onTabSelected(value);
+                  }
                   this.renderTabs();
                 }}
               >