+++ packages/common/src/Components/TextInput.jsx
@@ -7,9 +7,20 @@ import { ErrorIcon } from '../Icons';
 import Tooltip from './Tooltip';
 import textInputStyles from '../../statics/styles/text-input.scss';
 
-class TextInput extends React.Component {
+export default class TextInput extends React.Component {
+  static propTypes = {
+    inputRef: PropTypes.func,
+    theme: PropTypes.object.isRequired,
+    error: PropTypes.string,
+    showTooltip: PropTypes.bool,
+  };
+
+  static defaultProps = {
+    showTooltip: true,
+  };
+
   render() {
-    const { inputRef, error, theme, ...otherProps } = this.props;
+    const { inputRef, error, theme, showTooltip, ...otherProps } = this.props;
     const styles = mergeStyles({ styles: textInputStyles, theme });
     return (
       <div className={styles.textInput}>
@@ -20,26 +31,21 @@ class TextInput extends React.Component {
           })}
           {...otherProps}
         />
-        {error && (
-          <Tooltip
-            shouldRebuildOnUpdate={() => !!error}
-            content={error}
-            theme={theme}
-            moveBy={{ y: 0 }}
-            type={'error'}
-          >
+        {error &&
+          (showTooltip ? (
+            <Tooltip
+              shouldRebuildOnUpdate={() => !!error}
+              content={error}
+              theme={theme}
+              moveBy={{ y: 0 }}
+              type={'error'}
+            >
+              <ErrorIcon className={styles.textInput_errorIcon} />
+            </Tooltip>
+          ) : (
             <ErrorIcon className={styles.textInput_errorIcon} />
-          </Tooltip>
-        )}
+          ))}
       </div>
     );
   }
 }
-
-TextInput.propTypes = {
-  inputRef: PropTypes.func,
-  theme: PropTypes.object.isRequired,
-  error: PropTypes.string,
-};
-
-export default TextInput;