+++ packages/editor/src/RichContentEditor/RichContentEditor.jsx
@@ -43,16 +43,33 @@ class RichContentEditor extends Component {
       props.config.uiSettings || {}
     );
 
+    this.initContext();
     this.initPlugins();
   }
 
+  getEditorState = () => this.state.editorState;
+
   setEditorState = editorState => this.setState({ editorState });
+
+  initContext = () => {
+    const { theme, t, locale, anchorTarget, relValue, helpers, config, isMobile } = this.props;
+    this.contextualData = {
+      theme,
+      t,
+      locale,
+      anchorTarget,
+      relValue,
+      helpers,
+      config,
+      isMobile,
+      setEditorState: this.setEditorState,
+    };
+  };
 
   initPlugins() {
     const { helpers, plugins, config, isMobile, anchorTarget, relValue, t } = this.props;
 
     const { theme } = this.state;
-    const getEditorState = () => this.state.editorState;
     const { pluginInstances, pluginButtons, pluginTextButtons, pubsubs } = createPlugins({
       plugins,
       config,
@@ -62,7 +79,7 @@ class RichContentEditor extends Component {
       isMobile,
       anchorTarget,
       relValue,
-      getEditorState,
+      getEditorState: this.getEditorState,
       setEditorState: this.setEditorState,
     });
     this.initEditorToolbars(pluginButtons, pluginTextButtons);
@@ -94,8 +111,8 @@ class RichContentEditor extends Component {
       textToolbarType,
       textAlignment,
       theme: theme || {},
-      getEditorState: () => this.state.editorState,
-      setEditorState: editorState => this.setState({ editorState }),
+      getEditorState: this.getEditorState,
+      setEditorState: this.setEditorState,
       t,
       refId: this.refId,
       getToolbarSettings: config.getToolbarSettings,
@@ -156,7 +173,7 @@ class RichContentEditor extends Component {
   getEditorState = () => this.state.editorState;
 
   updateEditorState = editorState => {
-    this.setState({ editorState });
+    this.setEditorState(editorState);
     this.props.onChange && this.props.onChange(editorState);
   };
 
@@ -312,7 +329,7 @@ class RichContentEditor extends Component {
       [theme.desktop]: !isMobile && theme && theme.desktop,
     });
     return (
-      <Context.Provider theme={theme} {...this.props}>
+      <Context.Provider value={this.contextualData}>
         <Measure bounds onResize={({ bounds }) => this.updateBounds(bounds)}>
           {({ measureRef }) => (
             <div style={this.props.style} ref={measureRef} className={wrapperClassName}>
@@ -367,6 +384,7 @@ RichContentEditor.propTypes = {
   handleBeforeInput: PropTypes.func,
   handlePastedText: PropTypes.func,
   handleReturn: PropTypes.func,
+  locale: PropTypes.string,
 };
 
 RichContentEditor.defaultProps = {